Move tags to extras
