+++ extras/api/views.py
@@ -1,4 +1,5 @@
 import pyixapi
+from django.db.models import Count
 from drf_spectacular.types import OpenApiTypes
 from drf_spectacular.utils import OpenApiResponse, extend_schema
 from rest_framework import status
@@ -12,6 +13,7 @@ from extras.filters import (
     ConfigContextAssignmentFilterSet,
     ConfigContextFilterSet,
     ExportTemplateFilterSet,
+    TagFilterSet,
     WebhookFilterSet,
 )
 from extras.jobs import render_export_template
@@ -20,12 +22,10 @@ from extras.models import (
     ConfigContext,
     ConfigContextAssignment,
     ExportTemplate,
+    Tag,
     Webhook,
 )
-from peering_manager.api.viewsets import (
-    PeeringManagerModelViewSet,
-    PeeringManagerReadOnlyModelViewSet,
-)
+from peering_manager.api.viewsets import PeeringManagerModelViewSet
 
 from .serializers import (
     ConfigContextAssignmentSerializer,
@@ -33,6 +33,7 @@ from .serializers import (
     ExportTemplateSerializer,
     IXAPIAccountSerializer,
     IXAPISerializer,
+    TagSerializer,
     WebhookSerializer,
 )
 
@@ -152,6 +153,14 @@ class IXAPIViewSet(PeeringManagerModelViewSet):
         api.authenticate()
 
         return Response(data=api.customers.all())
+
+
+class TagViewSet(PeeringManagerModelViewSet):
+    queryset = Tag.objects.annotate(
+        tagged_items=Count("extras_taggeditem_items", distinct=True)
+    )
+    serializer_class = TagSerializer
+    filterset_class = TagFilterSet
 
 
 class WebhookViewSet(PeeringManagerModelViewSet):
