+++ extras/models/__init__.py
@@ -1,11 +1,14 @@
 from .configcontext import ConfigContext, ConfigContextAssignment
 from .ix_api import IXAPI
 from .models import ExportTemplate, Webhook
+from .tags import *
 
 __all__ = (
     "ConfigContext",
     "ConfigContextAssignment",
     "ExportTemplate",
     "IXAPI",
+    "TaggedItem",
+    "Tags",
     "Webhook",
 )
