+++ extras/tests/test_api.py
@@ -10,6 +10,7 @@ from extras.models import (
     ConfigContext,
     ConfigContextAssignment,
     ExportTemplate,
+    Tag,
     Webhook,
 )
 from peering.models import AutonomousSystem
@@ -233,6 +234,27 @@ class IXAPITest(StandardAPITestCases.View):
                 **self.header
             )
             self.assertHttpStatus(response, status.HTTP_200_OK)
+
+
+class TagTest(StandardAPITestCases.View):
+    model = Tag
+    brief_fields = ["id", "url", "name", "slug", "color"]
+    create_data = [
+        {"name": "Test 4", "slug": "test-4"},
+        {"name": "Test 5", "slug": "test-5"},
+        {"name": "Test 6", "slug": "test-6"},
+    ]
+    bulk_update_data = {"color": "000000"}
+
+    @classmethod
+    def setUpTestData(cls):
+        Tag.objects.bulk_create(
+            [
+                Tag(name="Test 1", slug="test-1", color="333333"),
+                Tag(name="Test 2", slug="test-2", color="333333"),
+                Tag(name="Test 3", slug="test-3", color="333333"),
+            ]
+        )
 
 
 class WebhookTest(StandardAPITestCases.View):
