+++ mkdocs.yml
@@ -66,7 +66,7 @@ nav:
         - Extras:
             - Config Context: "models/extras/configcontext.md"
             - IX-API: "models/extras/ixapi.md"
-            - Tag: "models/utils/tag.md"
+            - Tag: "models/extras/tag.md"
             - Webhook: "models/extras/webhook.md"
         - Token: "models/users/token.md"
     - Third-Party: "third-party.md"
