+++ templates/extras/tag/list.html
@@ -1,8 +1,8 @@
 {% extends '_base.html' %}
 {% load render_table from django_tables2 %}
 {% block actions %}
-{% if perms.utils.add_tag %}
-<a href="{% url 'utils:tag_add' %}" class="btn btn-primary">
+{% if perms.extras.add_tag %}
+<a href="{% url 'extras:tag_add' %}" class="btn btn-primary">
   <i class="fas fa-plus"></i> Add
 </a>
 {% endif %}
@@ -11,7 +11,7 @@
 {% block content %}
       <div class="row">
         <div class="col-md-9">
-          {% include 'generic/object_list.html' with bulk_edit_url='utils:tag_bulk_edit' bulk_delete_url='utils:tag_bulk_delete' %}
+          {% include 'generic/object_list.html' with bulk_edit_url='extras:tag_bulk_edit' bulk_delete_url='extras:tag_bulk_delete' %}
         </div>
         <div class="col-md-3">
           {% include 'utils/search_form.html' %}
