+++ templates/extras/tag/view.html
@@ -2,19 +2,19 @@
 {% load render_table from django_tables2 %}
 {% load helpers %}
 {% block actions %}
-{% if perms.utils.change_tag %}
-<a href="{% url 'utils:tag_edit' pk=instance.pk %}" class="btn btn-warning">
+{% if perms.extras.change_tag %}
+<a href="{% url 'extras:tag_edit' pk=instance.pk %}" class="btn btn-warning">
   <i class="fas fa-edit"></i> Edit
 </a>
 {% endif %}
 {% if perms.utils.delete_tag %}
-<a href="{% url 'utils:tag_delete' pk=instance.pk %}" class="btn btn-danger">
+<a href="{% url 'extras:tag_delete' pk=instance.pk %}" class="btn btn-danger">
   <i class="fas fa-trash"></i> Delete
 </a>
 {% endif %}
 {% endblock %}
 {% block breadcrumb %}
-<li class="breadcrumb-item"><a href="{% url 'utils:tag_list' %}">Tags</a></li>
+<li class="breadcrumb-item"><a href="{% url 'extras:tag_list' %}">Tags</a></li>
 <li class="breadcrumb-item active" aria-current="page">{% block title %}{{ instance }}{% endblock %}</li>
 {% endblock %}
 {% block content %}
