+++ utils/api/urls.py
@@ -6,7 +6,6 @@ router = PeeringManagerRouter()
 router.APIRootView = views.UtilsRootView
 
 router.register("object-changes", views.ObjectChangeViewSet)
-router.register("tags", views.TagViewSet)
 
 app_name = "utils-api"
 urlpatterns = router.urls
