+++ utils/tests/test_views.py
@@ -2,8 +2,9 @@ import uuid
 
 from django.contrib.auth.models import User
 
+from extras.models import Tag
 from utils.enums import ObjectChangeAction
-from utils.models import ObjectChange, Tag
+from utils.models import ObjectChange
 from utils.testing import ViewTestCases
 
 
@@ -29,28 +30,3 @@ class ObjectChangeTestCase(ViewTestCases.ReadOnlyObjectViewTestCase):
             change.user = user
             change.request_id = uid
             change.save()
-
-
-class TagTestCase(ViewTestCases.OrganizationalObjectViewTestCase):
-    model = Tag
-
-    test_changelog_object = None
-
-    @classmethod
-    def setUpTestData(cls):
-        Tag.objects.bulk_create(
-            (
-                Tag(name="Tag 1", slug="tag-1"),
-                Tag(name="Tag 2", slug="tag-2"),
-                Tag(name="Tag 3", slug="tag-3"),
-            )
-        )
-
-        cls.form_data = {
-            "name": "Tag 4",
-            "slug": "tag-4",
-            "color": "c0c0c0",
-            "comments": "Some comments",
-        }
-
-        cls.bulk_edit_data = {"color": "00ff00"}
